Add customized MaskRCNN
